Load Cart Items, Delete Cart Items

+++ app/src/main/java/com/example/foodorder/Database/CartDAO.java
@@ -22,7 +22,7 @@ public interface CartDAO {
     Single<Integer> countItemInCart(String uid);
 
     @Query("SELECT SUM((foodPrice+foodExtraPrice) * foodQuantity) FROM Cart WHERE uid=:uid")
-    Single<Double> sumPriceInCart(String uid);
+    Single<Long> sumPriceInCart(String uid);
 
     @Query("SELECT * FROM Cart WHERE foodId=:foodId AND uid=:uid")
     Single<CartItem> getItemInCart(String foodId, String uid);
