+++ app/src/main/java/com/example/foodorder/Database/CartDataSource.java
@@ -11,7 +11,9 @@ public interface CartDataSource {
 
     Single<Integer> countItemInCart(String uid);
 
-    Single<Double> sumPriceInCart(String uid);
+//    Single<Double> sumPriceInCart(String uid);
+    Single<Long> sumPriceInCart(String uid);
+
 
     Single<CartItem> getItemInCart(String foodId, String uid);
